backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# AI and ML
google-generativeai==0.3.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.0

# Redis and background jobs
redis==5.0.1
celery==5.3.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.25.2
PyPDF2==3.0.1
python-docx==1.1.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
